Move login state to separate component
